refactor(appcheck): Remove ktx imports
